make ProfileController

+++ demo_EMS/src/main/java/com/EMS_project/demo_EMS/Model/UserAuth.java
@@ -5,12 +5,13 @@ import jakarta.persistence.*;
 @Entity
 @Table(name = "user_auth")
 public class UserAuth {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(name = "user_id", nullable = false)
-    private Integer userId;
+    private Long userId;
 
     @Column(name = "user_type", nullable = false)
     @Enumerated(EnumType.STRING)
@@ -25,19 +26,19 @@ public class UserAuth {
     }
 
     // Getters and Setters
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
